refactor: 게시물 작성/수정 함께관리
